+++ HISTORY.md
@@ -36,6 +36,7 @@
 ### 2022-09-26
 * bitwise operators library: fixed #3118
 * updated German translation (removed obsolete argument-parameter mismatch error message)
+* objects: added Sprite>>hasPrimitiveCategories() method
 
 ### 2022-09-23
 * extensions: fixed palette category ghosting update when hiding / unhiding blocks